+++ Sources/PrintingKit/PrintItem.swift
@@ -13,26 +13,65 @@ import Foundation
  this library.
  
  The URL-based item types use optional URLs as a convenience,
- to make it easier to create item types. All printers in the
- library will throw a ``PrinterError`` if the URL is not set
- for the provided item.
+ to make it easier to create item types. A printer can throw
+ a ``PrinterError`` if the URL is not set for a certain item.
  */
 public enum PrintItem {
     
-    /// An JPG or PNG image at a certain URL.
-    case image(at: URL?)
+    /// An JPG or PNG image file at a certain URL.
+    case imageFile(at: URL?)
     
-    /// A PDF document at a certain URL.
-    case pdf(at: URL?)
+    /// JPG or PNG image data.
+    case imageData(Data)
+    
+    /// A PDF document file at a certain URL.
+    case pdfFile(at: URL?)
+    
+    /// PDF document data.
+    case pdfData(Data)
 }
 
-public extension PrintItem {
+extension Data {
     
-    /// The item's source URL, if any.
-    var sourceUrl: URL? {
-        switch self {
-        case .image(let url): return url
-        case .pdf(let url): return url
-        }
+    var canCreateExportFile: Bool {
+        fileManager.hasCachesDirectory
+    }
+    
+    var fileManager: FileManager {
+        .default
+    }
+    
+    func createExportFile(
+        withExtension ext: String
+    ) throws -> URL {
+        try fileManager.createCacheFile(
+            with: self,
+            fileExtension: ext
+        )
+    }
+}
+
+private extension FileManager {
+    
+    var cachesDirectoryUrl: URL? {
+        urls(for: .cachesDirectory, in: .userDomainMask).first
+    }
+    
+    var hasCachesDirectory: Bool {
+        cachesDirectoryUrl != nil
+    }
+    
+    func createCacheFile(
+        with data: Data,
+        fileExtension: String
+    ) throws -> URL {
+        let id = UUID().uuidString
+        guard let fileUrl = cachesDirectoryUrl?
+            .appendingPathComponent(id)
+            .appendingPathExtension(fileExtension)
+        else { throw PrinterError.cachesDirectoryDoesNotExist }
+        try? removeItem(at: fileUrl)
+        createFile(atPath: fileUrl.path, contents: data)
+        return fileUrl
     }
 }