+++ RELEASE_NOTES.md
@@ -8,11 +8,18 @@ Until then, deprecated features may be removed in the next minor version.
 
 ## 0.2
 
-This version adds PDF utilities. 
+This version adds PDF utilities and more print item types. 
+
+### ✨ New Features
 
 * `Pdf` is a new namespace with PDF-specific types.
 * `PdfDataSource` is a new protocol that is implemented by `NSAttributedString`.
+* `PrintItem` has new `.imageData` and `.pdfData` types.
 
+### 🚨 Breaking changes
+
+* `PrinterItem.image(at:)` has been renamed to `.imageFile`.
+* `PrinterItem.pdf(at:)` has been renamed to `.pdfFile`.
 
 
 ## 0.1